requirement.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
tmdbsimple==2.9.1
surprise==1.1.3
joblib==1.3.2
python-dotenv==1.0.0